WIP: 2c-d cleanup — 460 errors remaining (PromptAlertEvent unified across input/prompt_suggestions; EphemeralMessageModel unified between input/view; request_limit/requests_used return usize; CodebaseIndexSpeedbumpBannerState.id is InlineBannerId; server_conversation_token returns Option<ServerConversationToken>; .copied() on active_conversation_id callers)

+++ app/src/settings_view/billing_and_usage_page.rs
@@ -43,8 +43,8 @@ impl AIRequestUsageModel {
     pub fn ambient_only_credits_remaining(&self) -> Option<i64> { None }
     pub fn is_unlimited(&self) -> bool { false }
     pub fn next_refresh_time(&self) -> Option<chrono::DateTime<chrono::Utc>> { None }
-    pub fn request_limit(&self) -> i64 { 0 }
-    pub fn requests_used(&self) -> i64 { 0 }
+    pub fn request_limit(&self) -> usize { 0 }
+    pub fn requests_used(&self) -> usize { 0 }
     pub fn total_workspace_bonus_credits_remaining<W>(&self, _: W) -> i64 { 0 }
     pub fn total_current_workspace_bonus_credits_remaining<W>(&self, _: W) -> i64 { 0 }
     pub fn compute_buy_addon_credits_banner_display_state<W>(&self, _: W) -> crate::terminal::buy_credits_banner::BuyCreditsBannerDisplayState { crate::terminal::buy_credits_banner::BuyCreditsBannerDisplayState::Hidden }
